+++ FsInfoCat/FsInfoCat/IVolume.cs
@@ -16,7 +16,7 @@ namespace FsInfoCat
 
         bool? ReadOnly { get; set; }
 
-        long? MaxNameLength { get; set; }
+        int? MaxNameLength { get; set; }
 
         System.IO.DriveType Type { get; set; }
